backend/PythonAPI_Rev1/requirements.txt
pip~=23.2.1
SQLAlchemy~=2.0.39
fastapi~=0.115.11
pydantic~=2.10.6
python-dotenv~=1.0.1
pymysql~=1.1.1
uvicorn~=0.34.0